Split mixed test files so each maps to a single module
The webservice, tags and browser suites bundled tests for several
modules in one file, so a plain rename to test_init.py would have been
misleading. Split them by the module under test:

- test/webservice/test_webservice.py -> test_init.py (WebService, tasks,
  request queue, proxy, parser hook) + test_utils.py (picard.webservice.utils)
- test/tags/test_tags.py -> test_init.py (picard.tags helpers and ALL_TAGS
  validation) + test_tagvar.py (picard.tags.tagvar) + test_docs.py
  (picard.tags.docs rendering)
- test/browser/test_browser.py -> test_filelookup.py (picard.browser.filelookup)
  + test_server.py (picard.browser.server request handler and access log)

No test logic changed; the per-file test method counts sum to the same
totals as before the split (webservice 49, tags 37, browser 29).
