+++ test/tags/test_tagvar.py
@@ -25,30 +25,13 @@ import unittest.mock as mock
 from test.picardtestcase import PicardTestCase
 
 from picard.const import PICARD_URLS
-from picard.const.tags import ALL_TAGS
-from picard.options import (
-    Option,
-    get_option_title,
-)
+from picard.options import Option
 from picard.profile import profile_groups_add_setting
-from picard.tags import (
-    create_lang_desc_tag,
-    display_tag_name,
-    filterable_tag_names,
-    parse_lang_desc_tag,
-    preserved_tag_names,
-    script_variable_tag_names,
-    tag_names,
-)
-from picard.tags.docs import (
-    display_tag_full_description,
-    display_tag_tooltip,
-)
+from picard.tags import script_variable_tag_names
 from picard.tags.tagvar import (
     DocumentLink,
     TagVar,
     TagVars,
-    markdown,
 )
 
 
@@ -409,263 +392,3 @@ class TagVarsTest(PicardTestCase):
             '<p><strong>See Also:</strong> <a href="#_hidden_sd">%_hidden_sd%</a>; <a href="#sd_ld">%sd_ld%</a>.</p>'
         )
         self.assertEqual(tagvars.full_description_content(self.tagvar_everything), result)
-
-
-class UtilTagsTest(PicardTestCase):
-    def test_display_tag_name(self):
-        # Tag with no extra parts and no description
-        self.assertEqual(display_tag_name('tag'), 'tag')
-
-        # Tag with one extra part and no description
-        self.assertEqual(display_tag_name('tag:desc'), 'tag [desc]')
-
-        # Tag with blank extra part and no description
-        self.assertEqual(display_tag_name('tag:'), 'tag')
-
-        # Tag with multiple extra parts and no description
-        self.assertEqual(display_tag_name('tag:de:sc'), 'tag [de:sc]')
-
-        # Tag with no extra parts and short description
-        self.assertEqual(display_tag_name('originalyear'), 'Original Year')
-
-        # Tag with one extra part and short description
-        self.assertEqual(display_tag_name('originalyear:desc'), 'Original Year [desc]')
-
-        # Hidden tag with no extra parts and short description
-        self.assertEqual(display_tag_name('~length'), 'Length')
-
-        # Invalid hidden tag (not in ALL_TAGS)
-        self.assertEqual(display_tag_name('~lengthx'), '~lengthx')
-
-        # Empty tag
-        self.assertEqual(display_tag_name(''), '')
-
-    def test_parse_lang_desc_tag(self):
-        self.assertEqual(parse_lang_desc_tag('lyrics:eng'), ('eng', ''))
-        self.assertEqual(parse_lang_desc_tag('lyrics:XXX:foo'), ('XXX', 'foo'))
-        self.assertEqual(parse_lang_desc_tag('lyrics:XXX'), ('XXX', ''))
-        self.assertEqual(parse_lang_desc_tag('lyrics::foo', default_language='eng'), ('eng', 'foo'))
-
-    def test_parse_lang_desc_tag_invalid_language(self):
-        self.assertEqual(parse_lang_desc_tag('lyrics:de', default_language='eng'), ('eng', 'de'))
-        self.assertEqual(parse_lang_desc_tag('lyrics:toolong:foo', default_language='eng'), ('eng', 'toolong:foo'))
-
-    def test_parse_lang_desc_tag_description_with_colon(self):
-        # The description may itself contain colons and must be preserved in full.
-        self.assertEqual(parse_lang_desc_tag('lyrics:eng:a:b'), ('eng', 'a:b'))
-        self.assertEqual(parse_lang_desc_tag('lyrics::a:b', default_language='eng'), ('eng', 'a:b'))
-        self.assertEqual(parse_lang_desc_tag('comment:deu:foo:bar:baz'), ('deu', 'foo:bar:baz'))
-
-    def test_create_lang_desc_tag(self):
-        self.assertEqual('comment', create_lang_desc_tag('comment'))
-        self.assertEqual('comment:eng', create_lang_desc_tag('comment', language='eng'))
-        self.assertEqual('comment::foo', create_lang_desc_tag('comment', description='foo'))
-        self.assertEqual('lyrics:jpn:foo', create_lang_desc_tag('lyrics', language='jpn', description='foo'))
-        self.assertEqual(
-            'lyrics::foo', create_lang_desc_tag('lyrics', language='jpn', description='foo', default_language='jpn')
-        )
-
-    def test_display_tag_tooltip(self):
-        # Unknown tag
-        self.assertEqual(
-            display_tag_tooltip('unknown_test_variable'),
-            '<p><em>%unknown_test_variable%</em></p><p>No description available.</p>',
-        )
-
-        # Normal tag without notes.
-        self.assertEqual(display_tag_tooltip('album'), '<p><em>%album%</em></p><p>The title of the release.</p>')
-
-        # Normal tag without notes.
-        self.assertEqual(
-            display_tag_tooltip('_albumartists_sort'),
-            (
-                '<p><em>%_albumartists_sort%</em></p><p>The sort names of the album&#x27;s artists.</p><p><strong>Notes:'
-                '</strong> multi-value variable.</p>'
-            ),
-        )
-
-        # Normal tag with notes.
-        result = (
-            '<p><em>%albumsort%</em></p><p>The sort name of the title of the release.</p>'
-            '<p><strong>Notes:</strong> not provided from MusicBrainz data.</p>'
-        )
-        self.assertEqual(display_tag_tooltip('albumsort'), result)
-
-        # Hidden tag with notes, testing both prefixes '~' and '_'.
-        result = (
-            '<p><em>%_bitrate%</em></p><p>Approximate bitrate in kbps.</p>'
-            '<p><strong>Notes:</strong> preserved read-only; info from audio file; not provided from MusicBrainz data.</p>'
-        )
-        self.assertEqual(display_tag_tooltip('_bitrate'), result)
-        self.assertEqual(display_tag_tooltip('~bitrate'), result)
-
-        result = (
-            (
-                '<p><em>%performer%</em></p><p>The names of the performers for the specified type. These types include:</p>\n'
-                '<ul>\n'
-                '<li>vocals or instruments for the associated release or recording, where &quot;type&quot; can be &quot;<em>vocal</em>&quot;, '
-                '&quot;<em>guest guitar</em>&quot;, &quot;<em>solo violin</em>&quot;, etc.</li>\n'
-                '<li>the orchestra for the associated release or recording, where &quot;type&quot; is &quot;<em>orchestra</em>&quot;</li>\n'
-                '<li>the concert master for the associated release or recording, where &quot;type&quot; is &quot;<em>concertmaster</em>&quot;</li>\n'
-                '</ul><p><strong>Notes:</strong> multi-value variable.</p>'
-            )
-            if markdown is not None
-            else (
-                '<p><em>%performer%</em></p><p>The names of the performers for the specified type. These types include:'
-                '<br /><br />'
-                '- vocals or instruments for the associated release or recording, where &quot;type&quot; can be &quot;*vocal*&quot;, &quot;*guest '
-                'guitar*&quot;, &quot;*solo violin*&quot;, etc.<br />'
-                '- the orchestra for the associated release or recording, where &quot;type&quot; is &quot;*orchestra*&quot;<br />'
-                '- the concert master for the associated release or recording, where &quot;type&quot; is &quot;*concertmaster*&quot;</p>'
-                '<p><strong>Notes:</strong> multi-value variable.</p>'
-            )
-        )
-        self.assertEqual(display_tag_tooltip('performer'), result)
-
-    def test_display_tag_full_description(self):
-        item = TagVar(name='my_var', longdesc='The tag description.')
-
-        # Tag with description only
-        expected = "<p>The tag description.</p>"
-        self.assertEqual(display_tag_full_description(item), expected)
-
-        # Multi-value tag with description only
-        item.is_multi_value = True
-        expected = "<p>The tag description.</p><p><strong>Notes:</strong> multi-value variable.</p>"
-        self.assertEqual(display_tag_full_description(item), expected)
-        item.is_multi_value = False
-
-        # Tag with option setting only
-        Option('setting', 'my_script_var_opt', None, title='Option description')
-        item.related_options = ('my_script_var_opt',)
-        expected = "<p>The tag description.</p><p><strong>Option Settings:</strong> Option description.</p>"
-        self.assertEqual(display_tag_full_description(item), expected)
-        item.related_options = None
-
-        # Tag with plugin info
-        item.plugin_name = "My Plugin"
-        expected = "<p>The tag description.</p><p><strong>Plugin:</strong> My Plugin.</p>"
-        self.assertEqual(display_tag_full_description(item), expected)
-        item.plugin_name = None
-
-        # Tag with doc links
-        item.doc_links = (
-            DocumentLink('Link 1', 'https://example.com'),
-            DocumentLink('Link 2', 'https://example.com/foo'),
-        )
-        expected = "<p>The tag description.</p><p><strong>Links:</strong> <a href='https://example.com'>Link 1</a>; <a href='https://example.com/foo'>Link 2</a>.</p>"
-        self.assertEqual(display_tag_full_description(item), expected)
-        item.doc_links = None
-
-        # Tag with see also references
-        item.see_also = ('artist', 'albumartist')
-        expected = '<p>The tag description.</p><p><strong>See Also:</strong> <a href="#artist">%artist%</a>; <a href="#albumartist">%albumartist%</a>.</p>'
-        self.assertEqual(display_tag_full_description(item), expected)
-        item.see_also = None
-
-        # Tag with complex markdown (list items) and notes.
-        item.is_hidden = True
-        item.is_multi_value = True
-        item._longdesc = """Description of **My Var**:
-
-- Foo
-- Bar
-        """
-        result = (
-            (
-                '<p>Description of <strong>My Var</strong>:</p>\n'
-                '<ul>\n'
-                '<li>Foo</li>\n'
-                '<li>Bar</li>\n'
-                '</ul><p><strong>Notes:</strong> multi-value variable.</p>'
-            )
-            if markdown is not None
-            else (
-                '<p>Description of **My Var**:<br /><br />'
-                '- Foo<br />'
-                '- Bar</p>'
-                '<p><strong>Notes:</strong> multi-value variable.</p>'
-            )
-        )
-        self.assertEqual(display_tag_full_description(item), result)
-        item.is_hidden = False
-        item.is_multi_value = False
-
-        # Tag with no description
-        item._longdesc = None
-        expected = '<p>my_var</p>'
-        self.assertEqual(display_tag_full_description(item), expected)
-
-
-class UtilTagsOptionsTest(PicardTestCase):
-    def test_options_exist(self):
-        """Ensure all related options actually exist in the option settings registry (Option.registry)
-        and have a title set.
-        """
-        for tv in ALL_TAGS:
-            if tv.related_options is None:
-                continue
-            for opt in tv.related_options:
-                with self.subTest(tag=str(tv), option=opt):
-                    title = get_option_title(opt)
-                    self.assertIsNotNone(title, f"Missing related option setting '{opt}' in '{str(tv)}'")
-                    self.assertFalse(
-                        title.startswith('No title for setting'),
-                        f"Missing title for option setting '{opt}' in '{str(tv)}'",
-                    )
-
-
-class UtilTagsSeeAlsoTest(PicardTestCase):
-    def test_see_alsos_exist(self):
-        """Ensure all `see_also` tags actually exist in the `ALL_TAGS` collection and that a tag's
-        `see_also` tags do not refer to itself.
-        """
-        for tv in ALL_TAGS:
-            if tv.see_also is None:
-                continue
-            for also in tv.see_also:
-                with self.subTest(tag=str(tv), see_also=also):
-                    name = ALL_TAGS.script_name_from_name(also)
-                    self.assertIsNotNone(name, f"Invalid see_also '{also}' in '{str(tv)}' tag")
-                    self.assertNotEqual(name, str(tv), f"Circular see_also reference in '{str(tv)}' tag")
-
-
-class UtilTagsLinksTest(PicardTestCase):
-    def test_links_completeness(self):
-        """Ensure all `doc_links` entries have both a title and a link."""
-        for tv in ALL_TAGS:
-            if tv.doc_links is None:
-                continue
-            for doc_link in tv.doc_links:
-                with self.subTest(tag=str(tv), link=doc_link.link):
-                    title = doc_link.title.strip()
-                    link = doc_link.link.strip()
-                    self.assertNotEqual(title, '', f"Invalid link (missing title) in '{str(tv)}' tag")
-                    self.assertNotEqual(link, '', f"Invalid link (missing URL) in '{str(tv)}' tag")
-
-
-class TagsGeneratorTest(PicardTestCase):
-    def test_all_tags(self):
-        tags = list(tag_names())
-        self.assertTrue(len(tags) > 0)
-
-        for tag in tags:
-            with self.subTest(tag=tag):
-                self.assertTrue(ALL_TAGS.tagvar_from_name(tag).is_tag)
-
-    def test_all_preserved_tags(self):
-        tags = list(preserved_tag_names())
-        self.assertTrue(len(tags) > 0)
-
-        for tag in tags:
-            with self.subTest(tag=tag):
-                self.assertTrue(ALL_TAGS.tagvar_from_name(tag).is_preserved)
-
-    def test_all_filterable_tags(self):
-        tags = list(filterable_tag_names())
-        # Guard against the loop below passing vacuously, as the sibling tests do
-        self.assertTrue(len(tags) > 0)
-
-        for tag in tags:
-            with self.subTest(tag=tag):
-                self.assertTrue(ALL_TAGS.tagvar_from_name(tag).is_filterable)
