+++ test/webservice/test_init.py
@@ -43,11 +43,6 @@ from picard.webservice import (
     WebService,
     WSRequest,
 )
-from picard.webservice.utils import (
-    host_port_to_url,
-    hostkey_from_url,
-    port_from_qurl,
-)
 
 
 PROXY_SETTINGS = {
@@ -588,51 +583,3 @@ class WSRequestTest(PicardTestCase):
         )
         expected = 'http://example.org/path?a=1&a=1%26&b=2%26&c=%26&d=%26'
         self.assertEqual(request.url().toString(), expected)
-
-
-class WebServiceUtilsTest(PicardTestCase):
-    def test_port_from_qurl_http(self):
-        self.assertEqual(port_from_qurl(QUrl('http://example.org')), 80)
-
-    def test_port_from_qurl_http_other(self):
-        self.assertEqual(port_from_qurl(QUrl('http://example.org:666')), 666)
-
-    def test_port_from_qurl_https(self):
-        self.assertEqual(port_from_qurl(QUrl('https://example.org')), 443)
-
-    def test_port_from_qurl_https_other(self):
-        self.assertEqual(port_from_qurl(QUrl('https://example.org:666')), 666)
-
-    def test_port_from_qurl_exception(self):
-        with self.assertRaises(AttributeError):
-            port_from_qurl('xxx')
-
-    def test_hostkey_from_qurl_http(self):
-        self.assertEqual(hostkey_from_url(QUrl('http://example.org')), ('example.org', 80))
-
-    def test_hostkey_from_url_https_other(self):
-        self.assertEqual(hostkey_from_url('https://example.org:666'), ('example.org', 666))
-
-    def test_host_port_to_url_http_80(self):
-        self.assertEqual(
-            host_port_to_url('example.org', 80).toString(),
-            'http://example.org',
-        )
-
-    def test_host_port_to_url_https_443(self):
-        self.assertEqual(
-            host_port_to_url('example.org', 443).toString(),
-            'https://example.org',
-        )
-
-    def test_host_port_to_url_https_scheme_80(self):
-        self.assertEqual(
-            host_port_to_url('example.org', 80, scheme='https').toString(),
-            'https://example.org:80',
-        )
-
-    def test_host_port_to_url_http_666_with_path(self):
-        self.assertEqual(
-            host_port_to_url('example.org', 666, path='/abc').toString(),
-            'http://example.org:666/abc',
-        )
